+++ src/derivable/derivable.spec.ts
@@ -1,71 +1,16 @@
 import { expect } from 'chai';
-import { fromJS, List, Seq } from 'immutable';
+import { fromJS } from 'immutable';
 import { spy } from 'sinon';
-import { isAtom, isConstant } from '../extras';
-import { addObserver } from '../tracking';
-import { Atom, atom } from './atom';
-import { constant } from './constant';
-import { Derivable } from './derivable';
-import { Derivation, derivation } from './derivation';
-import { Lens } from './lens';
+import { isAtom } from '../extras';
+import { Derivable, SettableDerivable } from './derivable';
+import { Derivation } from './derivation';
+import { atom, constant, derivation } from './factories';
+import { testAccessors } from './mixins/accessors.spec';
+import { testBooleanFuncs } from './mixins/boolean-funcs.spec';
+import { testPluck } from './mixins/pluck.spec';
 
 export function testDerivable(factory: <V>(value: V) => Derivable<V>) {
-    describe('#get', () => {
-        it('should return the current state', () => {
-            const value$ = factory(123);
-            expect(value$.get()).to.equal(123);
-
-            if (isAtom(value$)) {
-                value$.set(456);
-                expect(value$.get()).to.equal(456);
-            }
-        });
-
-        it(`should ${factory === constant ? 'not ' : ''}be recorded inside a derivation'`, () => {
-            const value$ = factory(123);
-            expect(value$.observers).to.be.empty;
-            const derived$ = value$.derive(value => value + 876);
-            expect(value$.observers).to.be.empty;
-
-            // Simulate being observed to force derived$ to go into connected state.
-            addObserver(derived$, {} as any);
-            derived$.get();
-
-            if (factory === constant) {
-                expect(value$.observers).to.be.empty;
-            } else {
-                expect(value$.observers).to.have.length(1);
-                expect(value$.observers[0]).to.equal(derived$);
-            }
-        });
-    });
-
-    describe('#value', () => {
-        const a$ = factory('a');
-
-        if (!isConstant(a$)) {
-            it('should call #get() when getting the #value property', () => {
-                const s = spy(a$, 'get');
-
-                // Use the getter
-                expect(a$.value).to.equal('a');
-
-                expect(s).to.have.been.calledOnce;
-            });
-        }
-
-        if (isAtom(a$)) {
-            beforeEach('reset a$', () => a$.set('a'));
-
-            it('should call #set() when setting the #value property', () => {
-                const s = spy(a$, 'set');
-
-                a$.value = 'b';
-
-                expect(s).to.have.been.calledOnce.and.calledWithExactly('b');
-            });
-        }
-    });
+    testAccessors(factory);
 
     describe('#derive', () => {
         const oneGigabyte = 1024 * 1024 * 1024;
@@ -145,168 +90,9 @@ export function testDerivable(factory: <V>(value: V) => Derivable<V>) {
         });
     });
 
-    context('(boolean functions)', () => {
-        const true$ = factory(true);
-        const false$ = factory(false);
-        const bool$ = atom(false);
+    testBooleanFuncs(factory);
 
-        beforeEach('reset the atom', () => {
-            bool$.set(false);
-        });
-
-        describe('#is', () => {
-            it('should report equality on values', () => {
-                const value$ = factory('value');
-                expect(value$.is('value').get()).to.be.true;
-                expect(value$.is('something else').get()).to.be.false;
-            });
-
-            it('should report equality on derivables', () => {
-                const value$ = factory('value');
-                const atom$ = atom('value');
-                const valueIsAtom$ = value$.is(atom$);
-                expect(valueIsAtom$.get()).to.be.true;
-                atom$.set('something else');
-                expect(valueIsAtom$.get()).to.be.false;
-            });
-
-            it('should use the utils.equals function', () => {
-                const a$ = factory(List.of(1, 2, 3));
-                const b$ = factory(Seq.of(1, 2, 3));
-                expect(a$.is(b$).get()).to.be.true;
-            });
-        });
-
-        describe('#or', () => {
-            const trueOrBool$ = true$.or(bool$);
-            const falseOrBool$ = false$.or(bool$);
-
-            it('should apply boolean OR on the two derivables', () => {
-                expect(trueOrBool$.get()).to.be.true;
-                expect(falseOrBool$.get()).to.be.false;
-                bool$.set(true);
-                expect(trueOrBool$.get()).to.be.true;
-                expect(falseOrBool$.get()).to.be.true;
-            });
-
-            it('should not observe the right operand when the left operand is truthy', () => {
-                const s = spy(bool$, 'get');
-                trueOrBool$.get();
-                expect(s).not.to.have.been.called;
-                falseOrBool$.get();
-                expect(s).to.have.been.calledOnce;
-                s.restore();
-            });
-        });
-
-        describe('#and', () => {
-            const trueAndBool$ = true$.and(bool$);
-            const falseAndBool$ = false$.and(bool$);
-
-            it('should apply boolean AND on the two derivables', () => {
-                expect(trueAndBool$.get()).to.be.false;
-                expect(falseAndBool$.get()).to.be.false;
-                bool$.set(true);
-                expect(trueAndBool$.get()).to.be.true;
-                expect(falseAndBool$.get()).to.be.false;
-            });
-
-            it('should not observe the right operand when the left operand is falsey', () => {
-                const s = spy(bool$, 'get');
-                falseAndBool$.get();
-                expect(s).not.to.have.been.called;
-                trueAndBool$.get();
-                expect(s).to.have.been.calledOnce;
-                s.restore();
-            });
-        });
-
-        describe('#not', () => {
-            it('should apply boolean NOT on the input derivable', () => {
-                expect(false$.not().get()).to.be.true;
-                expect(true$.not().get()).to.be.false;
-            });
-        });
-
-    });
-
-    describe('#pluck', () => {
-        it('should pluck using a string or derivable string', () => {
-            const obj$ = factory({ a: 'valueOfA', b: 'valueOfB' });
-            const a$ = obj$.pluck('a');
-            expect(a$.get()).to.equal('valueOfA');
-
-            const prop$ = atom('a');
-            const item$ = obj$.pluck(prop$);
-            expect(item$.get()).to.equal('valueOfA');
-            prop$.set('b');
-            expect(item$.get()).to.equal('valueOfB');
-        });
-
-        it('should pluck using a Derivable of string or number', () => {
-            const arr$ = factory(['first', 'second']);
-            const first$ = arr$.pluck(0);
-            expect(first$.get()).to.equal('first');
-
-            const prop$ = atom(0);
-            const item$ = arr$.pluck(prop$);
-            expect(item$.get()).to.equal('first');
-            prop$.set(1);
-            expect(item$.get()).to.equal('second');
-        });
-
-        it('should pluck immutable values', () => {
-            const value$ = factory(fromJS({ a: ['value1', 'value2'] }));
-            const a$ = value$.pluck('a');
-            expect(a$.get()).to.equal(Seq.of('value1', 'value2'));
-            const value1$ = a$.pluck(1);
-            expect(value1$.get()).to.equal('value2');
-        });
-
-        context('(lensed)', () => {
-            class MyClass { constructor(public key: string) { } }
-            const value$ = factory(new MyClass('value'));
-
-            if (isAtom(value$)) {
-                beforeEach('reset the atom', () => value$.set(new MyClass('value')));
-
-                const plucked$ = value$.pluck('key');
-
-                it('should produce an Atom if the base was also an Atom', () => {
-                    expect(plucked$).to.be.an.instanceof(Derivation);
-                    expect(plucked$).to.be.an.instanceof(Lens);
-                    expect(isAtom(plucked$)).to.be.true;
-                });
-
-                it('should produce a lens that can change a property (cloning the object)', () => {
-                    const oldInstance = value$.get();
-                    expect(plucked$.get()).to.equal('value');
-                    plucked$.set('another value');
-                    expect(plucked$.get()).to.equal('another value');
-                    expect(value$.get()).to.deep.equal({ key: 'another value' });
-                    expect(value$.get()).to.not.equal(oldInstance);
-                });
-
-                it('should produce a lens that can change a property on an instance of a class', () => {
-                    expect(value$.get()).to.be.an.instanceOf(MyClass);
-                    plucked$.set('other value');
-                    expect(value$.get()).to.be.an.instanceOf(MyClass);
-                });
-
-                it('should produce a lens that can change immutable values', () => {
-                    const immValue$ = factory(fromJS({ a: ['value1', 'value2'], b: ['value3', 'value4'] })) as Atom<any>;
-                    const immPlucked$ = immValue$.pluck('b').pluck(0);
-                    immPlucked$.set('replaced');
-                    expect(immValue$.get()).to.equal(fromJS({ a: ['value1', 'value2'], b: ['replaced', 'value4'] }));
-                });
-
-                it('should not allow readonly immutable values to be changed', () => {
-                    const immValue$ = factory(Seq.of(1, 2, 3)) as Atom<any>;
-                    expect(() => immValue$.pluck(0).set(123)).to.throw();
-                });
-            }
-        });
-    });
+    testPluck(factory);
 
     describe('#react', () => {
         const value$ = factory('the value');
@@ -468,7 +254,7 @@ export function testDerivable(factory: <V>(value: V) => Derivable<V>) {
     });
 }
 
-function resetAtomTo<V>(a$: Atom<V>, value: V) {
+function resetAtomTo<V>(a$: SettableDerivable<V>, value: V) {
     a$.observers.forEach(obs => obs.disconnect());
     a$.set(value);
 }
