+++ src/derivable/index.ts
@@ -1,12 +1,8 @@
-// TODO
-import './boolean-funcs';
-
-export { Atom } from './atom.interface';
+export { Atom } from './atom';
 export { Constant } from './constant';
 export { DataSource } from './data-source';
-export { Derivable } from './derivable';
+export { Derivable, SettableDerivable, BaseDerivable } from './derivable';
 export { LensDescriptor, MonoLensDescriptor } from './lens.interface';
-export { unpack } from './unpack';
 export { Lens } from './lens';
 export { Derivation } from './derivation';
 export * from './factories';
