+++ src/index.ts
@@ -1,10 +1,11 @@
 export {
-    atom, Atom, constant, Constant, DataSource, Derivable, derivation, Derivation,
-    lens, Lens, LensDescriptor, MonoLensDescriptor, unpack,
+    atom, constant, DataSource, Derivable, derivation,
+    lens, LensDescriptor, MonoLensDescriptor, SettableDerivable,
 } from './derivable';
 
 export {
-    and, firstNotNull, isAtom, isConstant, isDerivable, isDerivation, isLens, lift, or, scan, struct, template, wrapPreviousState,
+    and, firstNotNull, isAtom, isConstant, isDerivable, isDerivation, isLens, isSettableDerivable,
+    lift, or, scan, struct, template, wrapPreviousState,
 } from './extras';
 
 export {
@@ -16,5 +17,8 @@ export {
 } from './transaction';
 
 export {
-    clone, equals, isPlainObject, setDebugMode,
+    clone, equals, isPlainObject, setDebugMode, unpack,
 } from './utils';
+
+import * as _advanced from './derivable';
+export { _advanced };
