--- src/tracking/tracked-observable.ts
@@ -1,31 +1,0 @@
-import { uniqueId } from '../utils';
-import { AutoCacheable, TrackedObservable, TrackedObserver } from './tracking';
-
-export abstract class BaseTrackedObservable implements TrackedObservable, AutoCacheable {
-
-    /**
-     * The unique ID of this Derivable. Can be used to uniquely identify this Derivable.
-     */
-    readonly id = uniqueId();
-
-    /**
-     * @internal
-     * The observers of this Derivable, do not use this in application code.
-     */
-    readonly observers: TrackedObserver[] = [];
-
-    /**
-     * Sets this Derivable to autoCache mode. This will cache the value of this Derivable the first time {@link #get} is called every tick
-     * and release this cache some time after this tick. The value is still guaranteed to be up-to-date with respect to changes in any of
-     * its dependencies, by using the same mechanism that is used by a reactor. It has a setup cost comparable to starting a reactor every
-     * first time #get is called per tick. Starting a reactor on a Derivable with an active and up-to-date cache is cheap though.
-     */
-    autoCache() { return this; }
-
-    /**
-     * @internal
-     * The current version of the state. This number gets incremented every time the state changes. Setting the state to
-     * an immutable object that is structurally equal to the previous immutable object is not considered a state change.
-     */
-    abstract version: number;
-}
